[ntuple, io] pass paths, names of ntuples for merging to RNTuple side
We construct a vector of string pairs and pass it through to the
RNTuple::Merge function for merging. This implements a "legacy bridge"
between hadd/TFileMerger and the new RNTuple library.

We adhere to the Merge(TCollection*, TFileMergeInfo*) interface without
using either class in the implementation by carefully casting the
RNTuple information vector and output file string pointers.

An example of this technique is shown here, which passes UBSan, ASan,
and strict aliasing checks on gcc 10 and clang 10: https://godbolt.org/z/7ff3e8